Improve all thhe `IDV POI tests
