+++ cypress/e2e/wip/KYC/idvDobMismatch.cy.js
@@ -1,18 +1,19 @@
+import { generateRandomName } from '../../../support/helper/utility'
+
 describe('QATEST-23042 IDV DOB Mismatch by Smile Identity provider', () => {
   beforeEach(() => {
     cy.c_visitResponsive('/')
-    cy.c_createRealAccount('gh')
+    cy.c_createRealAccount('ke')
     cy.c_login()
-    cy.c_navigateToPoiResponsive('Ghana')
+    cy.findByTestId('dt_traders_hub').should('be.visible')
+    cy.c_navigateToPoiResponsive('Kenya')
   })
 
   it('Should return Date of birth mismatch', () => {
-    cy.get('select[name="document_type"]').select(
-      'Social Security and National Insurance Trust (SSNIT)'
-    )
-    cy.findByLabelText('Enter your SSNIT number').type('C000000000000')
+    cy.get('select[name="document_type"]').select('National ID Number')
+    cy.findByLabelText('Enter your document number').type('00000000')
     cy.findByTestId('first_name').clear().type('Doe Joe')
-    cy.findByTestId('last_name').clear().type('Leo')
+    cy.findByTestId('last_name').clear().type(`Leo ${generateRandomName()}`)
     cy.findByTestId('date_of_birth').type('1991-08-23')
 
     cy.findByRole('button', { name: 'Verify' }).should('be.disabled')
@@ -29,5 +30,10 @@ describe('QATEST-23042 IDV DOB Mismatch by Smile Identity provider', () => {
     cy.findByText('Your identity verification failed because:').should(
       'be.visible'
     )
+    cy.findByTestId('DobMismatch')
+      .contains(
+        "The date of birth on your identity document doesn't match your profile."
+      )
+      .should('be.visible')
   })
 })
