+++ cypress/e2e/wip/KYC/idvDocumentRejected.cy.js
@@ -1,16 +1,19 @@
+import { generateRandomName } from '../../../support/helper/utility'
+
 describe('QATEST-22853 IDV Document Rejected by Smile Identity provider', () => {
   beforeEach(() => {
     cy.c_visitResponsive('/')
     cy.c_createRealAccount('gh')
     cy.c_login()
+    cy.findByTestId('dt_traders_hub').should('be.visible')
     cy.c_navigateToPoiResponsive('Ghana')
   })
 
   it('Should return Document Rejected', () => {
     cy.get('select[name="document_type"]').select('Passport')
     cy.findByLabelText('Enter your document number').type('G0000001')
-    cy.findByTestId('first_name').clear().type('adele')
-    cy.findByTestId('last_name').clear().type('Jojo')
+    cy.findByTestId('first_name').clear().type('John')
+    cy.findByTestId('last_name').clear().type(`Doe ${generateRandomName()}`)
     cy.findByTestId('date_of_birth').type('2000-09-20')
 
     cy.findByRole('button', { name: 'Verify' }).should('be.disabled')
