+++ cypress/e2e/wip/KYC/idvExpired.cy.js
@@ -1,8 +1,11 @@
+import { generateRandomName } from '../../../support/helper/utility'
+
 describe('QATEST-22808 IDV Expired scenario', () => {
   beforeEach(() => {
     cy.c_visitResponsive('/')
     cy.c_createRealAccount('ke')
     cy.c_login()
+    cy.findByTestId('dt_traders_hub').should('be.visible')
     cy.c_navigateToPoiResponsive('Kenya')
   })
 
@@ -10,7 +13,7 @@ describe('QATEST-22808 IDV Expired scenario', () => {
     cy.get('select[name="document_type"]').select('Passport')
     cy.findByLabelText('Enter your document number').type('A00000000')
     cy.findByTestId('first_name').clear().type('Joe Doe')
-    cy.findByTestId('last_name').clear().type('Leo')
+    cy.findByTestId('last_name').clear().type(`Leo ${generateRandomName()}`)
     cy.findByTestId('date_of_birth').type('2000-09-20')
     cy.get('.dc-checkbox__box').click()
     cy.findByRole('button', { name: 'Verify' }).click()
